app/requirements.txt
fastapi
uvicorn
greenlet
sqlalchemy
asyncpg
passlib[bcrypt]
python-jose
pandas